+++ lib/src/widget/circle_icon_button.dart
@@ -1,39 +1,40 @@
 import 'package:flutter/material.dart';
 
-class CircleIconButton extends StatelessWidget {
-  const CircleIconButton({
+class InstaPickerCircleIconButton extends StatelessWidget {
+  const InstaPickerCircleIconButton({
     super.key,
     required this.onTap,
     required this.icon,
     required this.theme,
+    required this.size,
   });
+
+  const InstaPickerCircleIconButton.unselectAll({
+    super.key,
+    required this.onTap,
+    required this.theme,
+    required this.size,
+  }) : icon = const Icon(Icons.layers_clear_sharp);
 
   final VoidCallback onTap;
   final Widget icon;
+  final double size;
   final ThemeData? theme;
 
   @override
   Widget build(BuildContext context) {
-    return ElevatedButton(
-      onPressed: onTap,
-      style: ElevatedButton.styleFrom(
-        fixedSize: const Size.fromRadius(14),
-        shape: CircleBorder(
-          side: theme?.iconTheme.color != null
-              ? BorderSide(
-                  color: theme!.iconTheme.color!.withOpacity(0.5),
-                  width: 0.1,
-                )
-              : BorderSide.none,
+    return SizedBox.square(
+      dimension: size,
+      child: ElevatedButton(
+        onPressed: onTap,
+        style: ElevatedButton.styleFrom(
+          shape: const CircleBorder(side: BorderSide.none),
+          padding: theme?.buttonTheme.padding,
+          backgroundColor: theme?.buttonTheme.colorScheme?.background,
+          foregroundColor: theme?.iconTheme.color,
         ),
-        padding: const EdgeInsets.all(6),
-        backgroundColor: theme?.buttonTheme.colorScheme?.background,
-        foregroundColor: theme?.iconTheme.color,
-        visualDensity: VisualDensity.compact,
-        tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-        textStyle: const TextStyle(fontSize: 12),
+        child: FittedBox(child: icon),
       ),
-      child: icon,
     );
   }
 }